Copyedit_datacenter
If it's possible to only use the registered trademark when Kafka is first mentioned on this page (in the first paragraph), that would be ideal. Currently, --8<-- "docs/shared/ccloud_setup.md" references the registered trademark as well. "Datacenter" has also been updated to be one word in the body text, but it'd be great to update the mention of this in the dropdown menu option in the left sidebar as well for consistency.