Copyedit_loyalty_rewards
For all the dog lovers out there who want to get really technical, "Caniche" seems to be the preferred term for "French Poodles," so I suggested updating the text to refer to "Poodles" for simplicity (I didn't make any changes to the code though, but it'd be great to make that consistent). I also recommended some changes in the event that some of the examples might be perceived differently than intended, just to be safe.